+++ src/browser_actions/browser-action.js
@@ -17,6 +17,22 @@ document.addEventListener('click', Sidebar.expandButtonListener);
 // List view controller
 // ////////////////////////////////////////////////////////////////
 
+function getBookmarkList(storage) {
+  const bookmarkList = Object.keys(storage)
+    .filter(key => (storage[key].type && storage[key].type === 'bookmark'))
+    .map(key => storage[key]);
+
+  bookmarkList.sort((a, b) => {
+    const refA = a.reference.toUpperCase();
+    const refB = b.reference.toUpperCase();
+    if (refA < refB) return -1;
+    if (refA > refB) return 1;
+    return 0;
+  });
+
+  return bookmarkList;
+}
+
 async function setListViewMode(viewMode, bookmarkList, shouldUpdateChart = true) {
   let readCount = 0;
 
@@ -43,7 +59,7 @@ async function setListViewMode(viewMode, bookmarkList, shouldUpdateChart = true)
       const card = fn(bookmark, manga);
       mangaListDom.appendChild(card);
 
-      if (card.classList.contains('flex-last')) readCount += 1;
+      if (card.classList.contains('order-last')) readCount += 1;
     });
 
     await Promise.all(promises);
@@ -57,11 +73,10 @@ async function setListViewMode(viewMode, bookmarkList, shouldUpdateChart = true)
 
 async function listViewModeListener(e) {
   try {
-    const storage = await browser.storage.sync.get(['view_mode', 'bookmark_list']);
+    const storage = await browser.storage.sync.get();
 
     // Initializes the view_mode and bookmark_list
     if (!storage.view_mode) storage.view_mode = {};
-    if (!storage.bookmark_list) storage.bookmark_list = [];
 
     storage.view_mode.list = e.target.id;
     await browser.storage.sync.set({ view_mode: storage.view_mode });
@@ -69,8 +84,11 @@ async function listViewModeListener(e) {
     // Clear previous list
     mangaListDom.innerHTML = '';
 
+    // Get the bookmarkList
+    const bookmarkList = getBookmarkList(storage);
+
     // Append new list
-    await setListViewMode(storage.view_mode.list, storage.bookmark_list, false);
+    await setListViewMode(storage.view_mode.list, bookmarkList, false);
 
     const container = document.getElementById('list-view-mode-container');
     const previousActive = container.getElementsByClassName('active')[0];
@@ -98,27 +116,28 @@ function updateCurrentChapter(bookmark) {
   const meta = mangaDom.getElementsByClassName('last-read-span')[0];
   meta.textContent = `Last read: ${moment(bookmark.last_read.date).format('LL')}`;
 
-  // Update List header
-  const chapterTracker = mangaDom.getElementsByClassName('list-chapter-tracker')[0];
-  if (chapterTracker) {
-    chapterTracker.innerText = chapterTracker.innerText.replace(/\d+\/(\d+)/, `${bookmark.last_read.chapter.index + 1}/$1`);
-  }
-
   // Update <select>
   const chapterSel = mangaDom.getElementsByTagName('select')[0];
   chapterSel.selectedIndex = (chapterSel.options.length - 1) - bookmark.last_read.chapter.index; // List is reversed
 
   const uptodate = bookmark.last_read.chapter.index === chapterSel.options.length - 1;
   if (uptodate) {
-    chapterSel.classList.replace('bg-danger', 'bg-success');
-    mangaDom.classList.add('flex-last');
+    chapterSel.classList.replace('bg-red', 'bg-green');
+    mangaDom.classList.add('order-last');
     Sidebar.updateChart(-1, 1);
+  }
+
+  // Update List header
+  const chapterTracker = mangaDom.getElementsByClassName('list-chapter-tracker')[0];
+  if (chapterTracker) {
+    chapterTracker.innerText = chapterTracker.innerText.replace(/\d+\/(\d+)/, `${bookmark.last_read.chapter.index + 1}/$1`);
+    if (uptodate) chapterTracker.classList.replace('text-danger', 'text-success');
   }
 }
 
 function updateChapterList(bookmark, chapterList) {
   const mangaDom = document.getElementById(`${bookmark.source}-${bookmark.reference}`);
-  const wasUptodate = mangaDom.classList.contains('flex-last');
+  const wasUptodate = mangaDom.classList.contains('order-last');
 
   // Update List header
   const chapterTracker = mangaDom.getElementsByClassName('list-chapter-tracker')[0];
@@ -129,7 +148,7 @@ function updateChapterList(bookmark, chapterList) {
   const chapterSel = mangaDom.getElementsByTagName('select')[0];
 
   if (wasUptodate) {
-    mangaDom.classList.remove('flex-last');
+    mangaDom.classList.remove('order-last');
     chapterSel.classList.replace('bg-success', 'bg-danger');
     Sidebar.updateChart(1, -1);
   }
@@ -212,9 +231,12 @@ window.onload = async () => {
       await browser.storage.sync.set({ badge_count: storage.badge_count });
     }
 
+    // Get the bookmarkList
+    const bookmarkList = getBookmarkList(storage);
+
     // Append manga list
-    if (storage.bookmark_list) {
-      await setListViewMode(storage.view_mode.list || 'list-rich', storage.bookmark_list);
+    if (bookmarkList) {
+      await setListViewMode(storage.view_mode.list || 'list-rich', bookmarkList);
     }
 
     // Add listener to list view radio button