+++ src/pages/util/export.js
@@ -1,7 +1,10 @@
+import { getError as FoxyError } from '../../util/foxyErrors';
+import Alert from '../../util/alerts';
+
 /**
  * Exports a backup file with all bookmarks.
  */
-export default async function exportBookmarks() {
+async function exportBookmarks() {
   // Create a hidden <a> tag
   const a = document.createElement('a');
   document.body.appendChild(a);
@@ -34,5 +37,19 @@ export default async function exportBookmarks() {
     window.URL.revokeObjectURL(url);
   } catch (err) {
     throw err;
+  }
+}
+
+/**
+ * Listens to the export button and generates the export file when clicked.
+ */
+export default async function exportButtonListener() {
+  try {
+    await exportBookmarks();
+  } catch (err) {
+    Alert('danger', {
+      title: browser.i18n.getMessage('alertErrorTitle'),
+      message: browser.i18n.getMessage('alertErrorMessage', [FoxyError().message, err.message]),
+    });
   }
 }