+++ src/util/http.js
@@ -1,4 +1,35 @@
 import { ErrorCode, getError as FoxyError } from './foxyErrors';
+import Cloudscrapper from './cloudscrapper';
+
+/**
+ * Fetches a new URL in order to solve a challenge (cloudflare ...).
+ * @param {string} url Required. String representing the URL to fetch (GET).
+ * @param {object} headers Required. The object with the headers to send.
+ * @returns {object} A promise which resolves to the HTMLDocument of the response body or its string representation.
+ */
+async function solveChallenge(url, headers) {
+  const options = {
+    headers,
+    credentials: 'include',
+  };
+
+  try {
+    const res = await window.fetch(url, options);
+    if (res.ok) {
+      const body = await res.text();
+
+      const resType = res.headers.get('Content-Type');
+      const parser = new DOMParser();
+      const parsedResponse = (resType.indexOf('text/html') !== -1) ? parser.parseFromString(body, 'text/html') : String(body);
+
+      return Promise.resolve(parsedResponse);
+    }
+
+    return Promise.reject(FoxyError(ErrorCode.SOURCE_SERVER_ERROR, `${res.status} ${res.statusText}`));
+  } catch (err) {
+    return Promise.reject(err);
+  }
+}
 
 /**
  * Uses the Fetch API to request information from an URL. Retries the fetch operation
@@ -19,7 +50,7 @@ export default async function fetch(url, onSuccess, options) {
   return new Promise((resolve, reject) => {
     const fetchAndRetry = async (n) => {
       try {
-        const res = await window.fetch(url);
+        const res = await window.fetch(url, { credentials: 'include' });
         if (res.ok) {
           const body = await res.text();
 
@@ -28,6 +59,19 @@ export default async function fetch(url, onSuccess, options) {
           const parsedResponse = (resType.indexOf('text/html') !== -1) ? parser.parseFromString(body, 'text/html') : String(body);
 
           resolve(onSuccess(parsedResponse));
+        } else if (res.status === 503 && res.headers.get('Server') === 'cloudflare') { // Cloudscrapper
+          const body = await res.text();
+          const answer = Cloudscrapper(res, body);
+
+          // Wait for 5sec before answering cloudflare challenge
+          setTimeout(async () => {
+            try {
+              const resBody = await solveChallenge(answer, { Referer: res.url });
+              resolve(onSuccess(resBody));
+            } catch (err) {
+              reject(FoxyError(ErrorCode.SOURCE_CLIENT_ERROR, `${res.status} ${res.statusText}`));
+            }
+          }, 5000);
         } else if (res.status >= 500) {
           throw FoxyError(ErrorCode.SOURCE_SERVER_ERROR, `${res.status} ${res.statusText}`);
         } else {